Fix projects grid, add shadow to the header
Fixed the projects grid by setting their max-height: fit content. That
seemed to work, even though the prokjects section is still bigger than
the right column.

Added shadow to the header. Tried initially to add it to the header
itself but it didn't work, same for the second-live div. Decided to work
in a round about way and added inset shadow to the main-content div;
that did the trick.
Also added some shadow to the project cards.